refactor(frontend): split PublicItemDetailPage god file into sections + hook
916-line god file split into modular structure under pages/public/:
- useRelatedItems.ts: 3 query strategies with fallback logic
- ItemInfoSection.tsx: specs, pricing, status badges
- ItemMediaSection.tsx: spinner 360 + image gallery + empty state
- RelatedItemsSection.tsx: related items display
- PublicItemDetailPage.tsx: layout skeleton (~241 lines)
- index.tsx: re-export for public/ directory

Original pages/PublicItemDetailPage.tsx reduced to a 2-line re-export
so App.tsx import is unchanged. Zero behavior changes.
